Second major rewrite
TokenStore now is an interface.
TokenSource is the only thing that handles token retrieval.
TokenStoreCookie is not implemented, and probably not going to be anytime soon.
Tests are mostly complete. Mostly.
Tests now require the --stderr flag to pass.
OauthConsumerBehavior should be functional but is not tested.
Schema and fixture have been updated.